% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_data_formats}
\alias{available_data_formats}
\title{List Available Data Formats}
\usage{
available_data_formats()
}
\value{
A named character vector of available data formats.
}
\description{
Scans the package source files for functions starting with "read_" to determine supported data formats.
}
