% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_report_templates}
\alias{available_report_templates}
\title{List Available Report Templates}
\usage{
available_report_templates()
}
\value{
A character vector of available report templates
}
\description{
Scans the package source files for available report templates to write to.
}
