% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_normalise.R
\name{batch_normalise}
\alias{batch_normalise}
\title{Batch Normalisation}
\usage{
batch_normalise(
  metaboprep,
  source_layer = "input",
  dest_layer = "batch_normalised"
)
}
\arguments{
\item{metaboprep}{an object of class Metaboprep}
}
\description{
Run batch normalisation based on the platform flag in the features data
}
