% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_normalization.R
\name{batch_normalization}
\alias{batch_normalization}
\title{A function to median normalize data sets by batch}
\usage{
batch_normalization(
  wdata,
  feature_data_sheet = NULL,
  sample_data_sheet = NULL,
  feature_runmode_col = NULL,
  batch_ids = NULL
)
}
\arguments{
\item{wdata}{the metabolite data frame samples in row, metabolites in columns}

\item{feature_data_sheet}{a data frame containing the feature annotation data}

\item{sample_data_sheet}{a data frame containing the sample annotation data}

\item{feature_runmode_col}{a string identifying the column name in the feature_data_sheet that identifies the run mode for each feature (metabolites of proteins).}

\item{batch_ids}{a string vector, with a length equal to the number of samples in the data set that identifies what batch each sample belongs to.}
}
\description{
This function median normalizes multivariable data such as metabolomic and proteomic data sets
}
\examples{
batch_normalization()
}
\keyword{batch}
\keyword{metabolomics}
\keyword{normalization}
\keyword{proteomics}
