% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_names}
\alias{clean_names}
\title{Standardize Column or Feature Names}
\usage{
clean_names(names)
}
\arguments{
\item{names}{`character vector` A vector of names to be cleaned.}
}
\value{
`character vector` A standardized version of the input names.
}
\description{
Cleans a character vector of names by replacing spaces with underscores,
removing special characters (`-`, `.`), replacing `%` with `pct`,
and converting to lowercase.
}
\examples{
clean_names(c("Sample ID", "Feature-Name.1", "Concentration \%"))
# Returns: c("sample_id", "featurename1", "concentration_pct")
}
