% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.power.binary.R
\name{eval.power.binary}
\alias{eval.power.binary}
\title{A Function to estimate power for a binary variable}
\usage{
eval.power.binary(N, effect, alpha)
}
\arguments{
\item{N}{Sample size}

\item{effect}{effect size}

\item{alpha}{significance level (Type 1 error)}
}
\description{
This function allows you estimate power for a binary variable given the sample size, effect size, TypeI error. Useful for QC Rmd Report.
}
\examples{
eval.power.binary()
}
\keyword{binary}
\keyword{estimator}
\keyword{power}
