% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous_power_plot.R
\name{eval.power.cont}
\alias{eval.power.cont}
\title{estimate power for continuous variable}
\usage{
eval.power.cont(N, n_coeff, effect, alpha)
}
\arguments{
\item{N}{Sample size}

\item{n_coeff}{degrees of freedom for numerator}

\item{effect}{effect size}

\item{alpha}{significance level (Type 1 error)}
}
\description{
This function estimates power for a continuous variable given the sample size, effect size, significance threshold, and the degrees of freedom.
}
\examples{
eval.power.cont(N = 1000, n_coeff = 1, effect = 0.0025, alpha = 0.05)

}
\keyword{continuous}
\keyword{power}
