% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_comets}
\alias{export_comets}
\title{Export Data to `COMETS` format}
\usage{
export_comets(metaboprep, directory, layer = NULL)
}
\arguments{
\item{metaboprep}{A `Metaboprep` object containing the data to be exported.}

\item{directory}{character, string specifying the path to the directory where the data should be written.}

\item{layer}{character, the name of the `metaboprep@data` layer (3rd array dimension) to write out}
}
\value{
the `Metaboprep` object, invisibly, for use in pipes
}
\description{
Export Data to `COMETS` format
}
