% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_data}
\alias{export_data}
\title{Export Data from a Metaboprep Object}
\usage{
export_data(metaboprep, directory)
}
\arguments{
\item{metaboprep}{A `Metaboprep` object containing the data to be exported.}

\item{directory}{A character string specifying the path to the directory where the data should be written.}
}
\value{
the `Metaboprep` object, invisibly, for use in pipes
}
\description{
Exports all data layers from a `Metaboprep` object to a structured directory format.
For each data layer, the function creates a subdirectory containing:
- the primary data matrix (`data.tsv`),
- associated feature and sample metadata (`features.tsv`, `samples.tsv`),
- feature and sample summaries (if present, `feature_summary.tsv`, `sample_summary.tsv`),
- a serialized feature tree (if present),
- and a `config.yml` file with additional metadata and processing parameters.
}
