% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_metaboprep}
\alias{export_metaboprep}
\title{Export Data to `Metaboprep` format}
\usage{
export_metaboprep(metaboprep, directory, ...)
}
\arguments{
\item{metaboprep}{A `Metaboprep` object containing the data to be exported.}

\item{directory}{character, string specifying the path to the directory where the data should be written.}

\item{...}{other parameters passed to \code{export_metaboprep()}, \code{export_comets()}, or \code{export_metaboanalyst()}.}
}
\value{
the `Metaboprep` object, invisibly, for use in pipes
}
\description{
Export Data to `Metaboprep` format
}
