% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.outliers.R
\name{feature.outliers}
\alias{feature.outliers}
\title{A Function to provide a count of the number outlying samples for a feature}
\usage{
feature.outliers(wdata, nsd = 5)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{nsd}{the number of standard deviation from the mean outliers are identified at. Default value is 5.}
}
\description{
This function takes a matrix of metaboite data (samples in rows, features in columns) and counts the number of outlying samples each feature has.
}
\examples{
feature.outliers()
}
\keyword{metabolomics}
