% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.tree.independence.R
\name{feature.tree.independence}
\alias{feature.tree.independence}
\title{A Function to identify independent features}
\usage{
feature.tree.independence(wdata)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}
}
\description{
This function allows you to identify independent features using Spearman's Rho, and a dendrogram tree cut step.
}
\examples{
feature.tree.independence()
}
\keyword{metabolomics}
