% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_describe.R
\name{feature_describe}
\alias{feature_describe}
\title{Summary Statistics for Features}
\usage{
feature_describe(data)
}
\arguments{
\item{data}{matrix, the metabolite data matrix. Samples in row, metabolites in columns}
}
\value{
a data frame of summary statistics for features (columns) of a matrix
}
\description{
This function allows you to 'describe' metabolite features using the describe() function from the psych package,
as well as estimate variance, a dispersion index, the coeficent of variation, and shapiro's W-statistic.
}
\examples{
ex_data = sapply(1:20, function(x){ rnorm(250, 40, 5) })
feature_describe(ex_data)

}
