% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalanced_power_plot.R
\name{find.PA.effect.sizes.2.sim}
\alias{find.PA.effect.sizes.2.sim}
\title{identify effect sizes}
\usage{
find.PA.effect.sizes.2.sim(mydata)
}
\arguments{
\item{mydata}{Your metabolite data matrix, with samples in rows}
}
\value{
a vector of effect sizes
}
\description{
This function estimates an appropriate distribution of effect sizes to simulate in a power analysis.
}
\examples{
ex_data = sapply(1:10, function(x){ rnorm(250, 40, 5) })
find.PA.effect.sizes.2.sim(ex_data)

}
\keyword{anlysis}
\keyword{power}
