% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.pairwise.n.filter.R
\name{greedy.pairwise.n.filter}
\alias{greedy.pairwise.n.filter}
\title{greedy selection}
\usage{
greedy.pairwise.n.filter(wdata, minN = 50)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{minN}{the minimum sample size (n) for pairwise comparisons}
}
\value{
a vector of feature names
}
\description{
This function identifies features that have less than a minimum number of complete pairwise observations and removes one of them, in a greedy fashion.
The need for this function is in instances where missingness is extreme between two features the number of paired observation between them may be to
to be informative. Thus one, but not both should be removed from the analysis to avoid analytical error based on sample sizes.
}
\examples{
set.seed(123)
ex_data = sapply(1:10, function(x){ rnorm(250, 40, 5) })
## define the data set
rownames(ex_data) = paste0("ind", 1:nrow(ex_data))
colnames(ex_data) = paste0("var", 1:ncol(ex_data))
## add in some missingness
ex_data[ sample(1:250, 200) ,1] = NA
ex_data[ sample(1:250, 200) ,2] = NA
ex_data[ sample(1:250, 200) ,3] = NA
## Estimate missingness and generate plots
greedy.pairwise.n.filter(ex_data)


}
\keyword{feature}
\keyword{selection}
