% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.outliers.R
\name{id.outliers}
\alias{id.outliers}
\title{identify outliers}
\usage{
id.outliers(x, nsd = 5)
}
\arguments{
\item{x}{a vector of numerical values}

\item{nsd}{the number of standard deviation from the mean outliers are identified at. Default value is 5.}
}
\value{
a vector indexing which samples are outliers
}
\description{
given a vector of data, identify those positions that are 'nsd' standard deviation units from the mean
}
\examples{
ex_data = rnbinom(500, mu = 40, size = 5)
id.outliers(ex_data, nsd = 2)

}
\keyword{outliers}
