% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbalanced_power_plot.R
\name{imbalanced_power_plot}
\alias{imbalanced_power_plot}
\title{binary trait imbalanced design power analysis plot}
\usage{
imbalanced_power_plot(mydata)
}
\arguments{
\item{mydata}{a numeric data matrix with samples in rows and features in columns}
}
\value{
a ggplot2 object
}
\description{
This function (1) estimates an informative distribution of effect and power estimates given your datas total sample size, over a distribution of imbalanced sample sizes and (2) returns a summary plot.
}
\examples{
ex_data = matrix(NA, 1000, 2)
imbalanced_power_plot( ex_data )

}
\keyword{analysis}
\keyword{binary}
\keyword{plot}
\keyword{power}
\keyword{trait}
