% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_impute.R
\name{median_impute}
\alias{median_impute}
\title{A Function to impute features (columns) of a metabolome matrix to median estimates}
\usage{
median_impute(wdata)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}
}
\description{
This function allows you  to impute features (columns) of a metabolome matrix to median estimates. Useful for PCA.
}
\examples{
median_impute()
}
\keyword{metabolomics}
