% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2batch.R
\name{met2batch}
\alias{met2batch}
\title{A Function to evaluate the effect of a laboratory batch variable on all metabolite feautures in a data matrix}
\usage{
met2batch(wdata, batch)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{batch}{a single vector containing a vector based batch variable}
}
\description{
This function allows you to estimate effects of laboratory batch variable on metabolite features.
}
\examples{
met2batch()
}
\keyword{metabolomics}
