% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metaboprep.R
\name{Metaboprep}
\alias{Metaboprep}
\title{Metaboprep Object}
\usage{
Metaboprep(data, samples, features)
}
\arguments{
\item{data}{numeric matrix, the data matrix containing metabolite values (not to be set directly).}

\item{samples}{data.frame, a data frame containing sample-related information (not to be set directly).}

\item{features}{data.frame, a data frame containing feature-related information (not to be set directly).}

\item{exclusions}{list, holds exclusion codes for data masking (not to be set directly).}

\item{feature_summary}{numeric matrix, summary statistics for features (not to be set directly).}

\item{sample_summary}{numeric matrix, summary statistics for samples (not to be set directly).}
}
\value{
An object of class `Metaboprep` (S7 class).
}
\description{
A `Metaboprep` object is a container for matrices of metabolite data, along with associated metadata.
It allows for efficient storage and manipulation of data, supporting quality control, transformations,
and various analyses. This object facilitates easy access to data layers, sample and feature summaries,
outlier treatment, and more.
}
\section{Slots}{

\describe{
\item{\code{data}}{numeric matrix, the metabolite data.}

\item{\code{samples}}{data.frame, the samples data frame}

\item{\code{features}}{data.frame, the features data frame}

\item{\code{exclusions}}{list, exclusion codes (mask for data).}

\item{\code{feature_summary}}{numeric matrix, feature summary statistics.}

\item{\code{sample_summary}}{numeric matrix, sample summary statistics.}
}}

