% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{missingness}
\alias{missingness}
\title{Estimate Missingness}
\usage{
missingness(data, by = "row")
}
\arguments{
\item{data}{matrix, a numeric matrix with samples in rows and features in columns}

\item{by}{character, whether to calculate missingness by rows (samples) or column (features)}
}
\value{
data.frame, a data frame of missingness estimates for each sample/feature.
}
\description{
This function estimates missingness in a matrix of data and provides an option to exclude certain columns or features from the analysis, such as xenobiotics (with high missingness rates) in metabolomics data sets.
}
