% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.anova.R
\name{multivariate.anova}
\alias{multivariate.anova}
\title{A Function to perform multivariate analysis over a dependent and an independent variable}
\usage{
multivariate.anova(dep, indep_df)
}
\arguments{
\item{dep}{a vector of the dependent variable}

\item{indep_df}{a data frame of the independent variable}
}
\description{
This function performs univariate analysis over a dependent and an independent variable
}
\examples{
multivariate.anova()
}
\keyword{metabolomics}
