% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_anova.R
\name{multivariate_anova}
\alias{multivariate_anova}
\title{multivariate analysis}
\usage{
multivariate_anova(dep, indep_df)
}
\arguments{
\item{dep}{a vector of the dependent variable}

\item{indep_df}{a data frame of the independent variable}
}
\value{
ggplot2 table figure of
}
\description{
This function performs a multivariate analysis over a dependent|response and numerous independent|explanatory variable
}
\examples{
## simulate some correlated data
set.seed(1110)
n <- 250
mu <- c(5, 45, 25)
cmat <- matrix(c(1, 0.5, 0.3,
                 0.5, 1, 0.25,
                 0.3, 0.25, 1), nrow = 3, byrow = TRUE)
L <- chol(cmat)
Z <- matrix(rnorm(n * 3), nrow = n)
ex_data <- Z \%*\% L
ex_data <- sweep(ex_data, 2, mu, "+")
colnames(ex_data) = c("outcome","age","bmi")
multivariate_anova(dep = ex_data[,1], indep_df = ex_data[, 2:3])

}
\keyword{ANOVA}
\keyword{metabolomics}
\keyword{multivariate}
