% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ng_anno}
\alias{ng_anno}
\title{Nightingale Health metabolomics annotation data set}
\format{
A data frame with 233 rows and 7 variables:
\describe{
  \item{metabolite}{metabolite id}
  \item{raw.label}{metabolite name and units}
  \item{class}{metabolite annotation class}
  \item{subclass}{metabolite annotation subclass}
  \item{label}{metabolite name and units}
  \item{label.no.units}{metabolite name without units}
  \item{derived_features}{a binary yes|no indicating if the metabolite variable is a variable derived of two or more other features in the data set}
  ...
}
}
\source{
\url{http://nightingalehealth.com/}
}
\usage{
ng_anno
}
\description{
A dataset containing annotation data on Nightingale Health NMR metabolites (lipids)
}
\keyword{datasets}
