% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{outliers}
\alias{outliers}
\title{Identify Outliers}
\usage{
outliers(x, nsd = 3)
}
\arguments{
\item{x}{a numerical vector of data}

\item{nsd}{the number of SD units from the mean to be used as an outlier cutoff.}
}
\value{
a list object of length three. (1) a vector of sample indexes indicating the outliers, (2) the lower outlier cuttoff value, (3) the upper outlier cuttoff value.
}
\description{
This function identifies outliers from a vector of data at SD units from the mean.
}
\examples{
ex_data = rnbinom(500, mu = 40, size = 5)
outliers(ex_data)

}
\keyword{outliers}
