% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcapairs_bymoose.R
\name{pcapairs_bymoose}
\alias{pcapairs_bymoose}
\title{pca pairs plot}
\usage{
pcapairs_bymoose(myloadings, varexp, pcol = "dodgerblue")
}
\arguments{
\item{myloadings}{a matrix or data frame of PC loadings (only those you would like to plot).}

\item{varexp}{a vector of the the variance explained by each PC}

\item{pcol}{plot colors for the dots background}
}
\value{
a base R plot
}
\description{
This function generates an upper triangle PCA plot for all pairs of PC loadings provided.
}
\examples{
ex_data = sapply(1:5, function(x){rnorm( 50, 0, 2) })
## add in some extreme values to a sample
ex_data[1,] = ex_data[1,] + sample( c(8, -8), ncol(ex_data), replace = TRUE )
ex_data[2,] = ex_data[2,] + sample( c(3, -3), ncol(ex_data), replace = TRUE )
## plot
pcapairs_bymoose(myloadings = ex_data, 
    varexp = rep(1/ncol(ex_data), ncol(ex_data)),
    pcol = "tomato" )

}
\keyword{PCA}
\keyword{metabolomics}
\keyword{pairs}
\keyword{plot}
