% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.metabolite.qc.R
\name{perform.metabolite.qc}
\alias{perform.metabolite.qc}
\title{A Super Function to perform all of your metabolomics QC}
\usage{
perform.metabolite.qc(wdata, fmis = 0.2, smis = 0.2,
  feature_colnames_2_exclude = NA, tpa_out_SD = 5, PC_out_SD = 5,
  ind_feature_names = NA)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{fmis}{defaulted at 0.2, this defines the feature missingness cutoff}

\item{smis}{defaulted at 0.2, this defines the sample missingness cutoff}

\item{feature_colnames_2_exclude}{names of columns to be excluded from this analysis, such as for Xenobiotics. Pass NA to this paramater to exclude this step.}

\item{tpa_out_SD}{defaulted at '5', this defines the number of standard deviation from the mean in which samples will be excluded for total peak area. Pass NA to this paramater to exclude this step.}

\item{PC_out_SD}{defaulted at '5', this defines the number of standard deviation from the mean in which samples will be excluded for principle components. NA is NOT an excepted paramater.}

\item{ind_feature_names}{If you have previously identified independent features in your data set they can be provided here a vector of column names. Pass NA to this paramater to have the function estimate indpendence for you.}
}
\description{
This function allows you perform the key quality controls steps on a metabolomics data set
}
\examples{
perform.metabolite.qc()
}
\keyword{metabolomics}
