% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.in.metabolon.R
\name{read.in.metabolon}
\alias{read.in.metabolon}
\title{A Function to read in Metabolon excel sheets}
\usage{
read.in.metabolon(file2process, data_dir, projectname)
}
\arguments{
\item{file2process}{the name of the xls file to process}

\item{data_dir}{the full path to the directory holding your Metabolon excel file}

\item{projectname}{a name for your project}
}
\description{
This function allows you to read in Metabolon raw data and push out flat text files
}
\examples{
read.in.metabolon()
}
\keyword{metabolon}
