% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metabolon.R
\name{read_metabolon}
\alias{read_metabolon}
\title{Read Metabolon Data
Reads and processes a commercial Metabolon Excel sheet in v1.1, v1.2, or v2 format. Returns either a structured `Metaboprep` object or a list of components.}
\usage{
read_metabolon(
  filepath,
  sheet = NULL,
  feature_sheet = NULL,
  feature_id_col = NULL,
  sample_sheet = NULL,
  sample_id_col = NULL,
  return_Metaboprep = TRUE
)
}
\arguments{
\item{filepath}{character, commercial Metabolon excel sheet with extension .xls or .xlsx}

\item{sheet}{character or integer, the excel sheet name (or index) from which to read.}

\item{feature_sheet}{character or integer, the excel sheet name (or index) from which to read the feature data.}

\item{feature_id_col}{character, the excel column containing the feature_id mapping to the data.}

\item{sample_sheet}{character or integer, the excel sheet name (or index) from which to read the sample data.}

\item{sample_id_col}{character, the excel column containing the sample_id mapping to the data.}

\item{return_Metaboprep}{logical, if TRUE (default) return a Metaboprep object, if FALSE return a list.}
}
\value{
Metaboprep object or list,  list(data = matrix, samples = samples data.frame, features = features data.frame)

If `return_Metaboprep = TRUE` (default), returns a `Metaboprep` object containing the input data.
If `return_Metaboprep = FALSE`, returns a named list with the following elements:
\describe{
  \item{data}{A numeric matrix of metabolite values with samples as rows and features as columns.}
  \item{samples}{A `data.frame` containing metadata for each sample (must include `sample_id`).}
  \item{features}{A `data.frame` containing metadata for each metabolite feature (must include `feature_id`).}
}
}
\description{
Read Metabolon Data
Reads and processes a commercial Metabolon Excel sheet in v1.1, v1.2, or v2 format. Returns either a structured `Metaboprep` object or a list of components.
}
\details{
This function reads a commercial Metabolon Excel file in either single-sheet (v1/v2) or multi-sheet (v3) format.
It extracts sample-level and feature-level metadata alongside the numeric data matrix. 
Users can choose to return either a structured `Metaboprep` S7 object or a basic list via the `return_Metaboprep` parameter.
}
\examples{
# version 1.1 data format
filepath1 <- system.file("extdata", "metabolon_v1.1_example.xlsx", package = "metaboprep")
m <- read_metabolon(filepath1, sheet = 2)

# version 1.2 data format (different column names)
filepath2 <- system.file("extdata", "metabolon_v1.2_example.xlsx", package = "metaboprep")
m <- read_metabolon(filepath2, sheet = 'OrigScale')


# version 2 data format
filepath3 <- system.file("extdata", "metabolon_v2_example.xlsx", package = "metaboprep")
m <- read_metabolon(filepath3, 
                    sheet = 'Batch-normalized Data', 
                    feature_sheet = 'Chemical Annotation', 
                    feature_id_col = 'CHEM_ID', 
                    sample_sheet = 'Sample Meta Data', 
                    sample_id_col = 'PARENT_SAMPLE_NAME')

}
