% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nightingale_v1.R
\name{read_nightingale_v1}
\alias{read_nightingale_v1}
\title{Read Nightingale Data (format 1)}
\usage{
read_nightingale_v1(filepath)
}
\arguments{
\item{filepath}{character, commercial Nightingale excel sheet with extension .xls or .xlsx}
}
\value{
list,  list(data = 3D matrix, samples = samples data.table, features = features data.table)
}
\description{
Read Nightingale Data (format 1)
}
