% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_olink.R
\name{read_olink_v1}
\alias{read_olink_v1}
\title{Read and Process Olink NPX Data File}
\usage{
read_olink_v1(filepath)
}
\arguments{
\item{filepath}{A string specifying the path to the Olink NPX file.}
}
\value{
A named list with the following elements:
\describe{
  \item{data}{A matrix of NPX values with `SampleID` as rows and `OlinkID` as columns, containing only sample data.}
  \item{samples}{A `data.frame` containing metadata for samples.}
  \item{features}{A `data.frame` containing feature-level metadata for samples.}
  \item{controls}{A matrix of NPX values for control samples.}
  \item{control_metadata}{A `data.frame` containing metadata for control samples.}
}
}
\description{
This function reads and processes an Olink NPX file in long format. It supports `.csv`, `.xls`, `.xlsx`, `.txt`, `.zip`, and `.parquet` formats, and returns a list of matrices and metadata frames for further analysis.
}
\details{
The function checks whether the input data is in long format by verifying the presence of duplicate `SampleID` values. It also accommodates two variants of Olink files:
\itemize{
  \item Files that include a `Sample_Type` column with values `"SAMPLE"` and `"CONTROL"`.
  \item Files that use the `SampleID` column to label control samples (e.g., entries containing `"CONTROL"`).
}

If neither format is detected, the function stops with an error indicating that the data is likely not from Olink.
}
\examples{
\dontrun{
  filepath <- system.file("extdata", "example_olink_data.txt", package = "metaboprep")
  olink_data <- read_olink_v1(filepath)
}
}
