% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rntransform.R
\name{rntransform}
\alias{rntransform}
\title{A copy of the rntransform function from the GenAbel package, edited for the option of spliting tied values.}
\usage{
rntransform(formula, data, family = gaussian, split_ties = FALSE)
}
\arguments{
\item{formula}{}

\item{data}{your vector of numerical values to rank normal transform}

\item{family}{Defaulted to gaussian. Do not suggest editing}

\item{split_ties}{Defualted to FALSE. If TRUE is provided all tied values will be randomly ranked in the function rank()}
}
\description{
This function allows you to rank normal transform a vector of data.
}
\examples{
rntransform()
}
\keyword{nigtingale}
