% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.pa.imbalanced.power.make.plot.R
\name{run.pa.imbalanced.power.make.plot}
\alias{run.pa.imbalanced.power.make.plot}
\title{A Function to simulate power for a spectrum of effect sizes, and generate plots, assuming a presence absence binary trait in an imbalanced design}
\usage{
run.pa.imbalanced.power.make.plot(mydata)
}
\arguments{
\item{mydata}{Your metabolite data matrix, with samples in rows}
}
\description{
This function allows simulate power for numberous effect estimates for both binary and continuous variables and generate summary plots. Useful for QC Rmd Report.
}
\examples{
run.pa.imbalanced.power.make.plot()
}
\keyword{metabolomics}
