% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.missingness.R
\name{sample.missingness}
\alias{sample.missingness}
\title{A Function to estimate sample missingness}
\usage{
sample.missingness(wdata, excludethesefeatures = NA)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{excludethesefeatures}{a vector of feature names (i.e. column names) to exclude from missingness estimates}
}
\description{
This function allows estiamte sample missingness in a matrix of data and provides an option to exclude certain columns or metabolite features from the analysis, such as xenobiotics
}
\examples{
sample.missingness()
}
\keyword{metabolomics}
