% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.outliers.R
\name{sample.outliers}
\alias{sample.outliers}
\title{A Function to provide a count of the number outlying features for a sample}
\usage{
sample.outliers(wdata, nsd = 5)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{nsd}{the number of standard deviation from the mean outliers are identified at. Default value is 5.}
}
\description{
This function takes a matrix of metaboite data (samples in rows, features in columns) and counts the number of outlying features each sample has.
}
\examples{
sample.outliers()
}
\keyword{metabolomics}
