% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.sum.stats.R
\name{sample.sum.stats}
\alias{sample.sum.stats}
\title{summary statistics for samples}
\usage{
sample.sum.stats(wdata, feature_names_2_exclude = NA)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{feature_names_2_exclude}{a vector of feature|column names to exclude from missingness estimates}
}
\value{
a data frame of summary statistics
}
\description{
This function estimates summary statistics for samples in a matrix of numeric features. This includes missingness, total peak area, and a count of the number of outlying features for a sample.
}
\examples{
sample.sum.stats()

}
\keyword{sample}
\keyword{statistics}
\keyword{summary}
