% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_summary.R
\name{sample_summary}
\alias{sample_summary}
\title{Sample Summary Statistics}
\usage{
sample_summary(
  metaboprep,
  source_layer = "input",
  outlier_udist = 5,
  exclude_derived = FALSE,
  exclude_xenobiotics = FALSE,
  sample_ids = NULL,
  feature_ids = NULL,
  output = "object"
)
}
\arguments{
\item{metaboprep}{an object of class Metaboprep}

\item{source_layer}{character, the data layer to summarise}

\item{outlier_udist}{the unit distance in SD or IQR from the mean or median estimate, respectively outliers are identified at. Default value is 5.}

\item{sample_ids}{character, vector of sample ids to work with}

\item{feature_ids}{character, vector of feature ids to work with}

\item{output}{character, type of output, either `object` to return the updated metaboprep object, or data.frame to return the data.}
}
\description{
Summarise the sample data
}
