% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{shiny_app}
\alias{shiny_app}
\title{Metaboprep Shiny App}
\usage{
shiny_app(metaboprep)
}
\arguments{
\item{metaboprep}{an object of class Metaboprep}
}
\value{
Runs a Shiny app
}
\description{
Launch a Shiny app to explore the Metaboprep object
}
