% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_peak_area.R
\name{total_peak_area}
\alias{total_peak_area}
\title{Estimates total peak abundance}
\usage{
total_peak_area(data, ztransform = TRUE)
}
\arguments{
\item{data}{matrix, the metabolite data matrix. Samples in rows, metabolites in columns}

\item{ztransform}{logical, should the feature data be z-transformed and absolute value minimum, mean shifted prior to summing the feature values. TRUE or FALSE.}
}
\value{
a data frame of estimates for (1) total peak abundance and (2) total peak abundance at complete features for each samples
}
\description{
This function estimates total peak abundance|area for numeric data in a matrix, for (1) all features and (2) all features with complete data.
}
