% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_and_independent_features.R
\name{tree_and_independent_features}
\alias{tree_and_independent_features}
\title{A Function to identify independent features}
\usage{
tree_and_independent_features(wdata, minimum_samplesize = 50,
  treecutheight = 0.2)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{minimum_samplesize}{the metabolite data matrix. samples in row, metabolites in columns}

\item{treecutheight}{the tree cut height. A value of 0.2 (1-Spearman's rho) is equivalent to saying that features with a rho >= 0.8 are NOT independent.}
}
\description{
This function allows you to identify independent features using Spearman's Rho, and a dendrogram tree cut step.
}
\examples{
tree_and_independent_features()
}
\keyword{metabolomics}
