% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.by.factor.R
\name{variable.by.factor}
\alias{variable.by.factor}
\title{ggplot2 violin plot}
\usage{
variable.by.factor(
  dep,
  indep,
  dep_name = "dependent",
  indep_name = "independent",
  orderfactor = TRUE,
  violin = TRUE
)
}
\arguments{
\item{dep}{a vector of the dependent variable}

\item{indep}{a vector of the independent variable}

\item{dep_name}{name of the dependent variable}

\item{indep_name}{name of the independent variable}

\item{orderfactor}{order factors alphebetically}

\item{violin}{box plot or violin plot. violin = TRUE is default}
}
\value{
a ggplot2 object
}
\description{
This function performs univariate linear analysis of a dependent and an independent variable and generates a viloin or box plot to illustrate the associated structure.
}
\examples{
x = c( rnorm(20, 10, 2), rnorm(20, 20, 2) )
y = as.factor( c( rep("A", 20), rep("B", 20)  ) )
variable.by.factor(dep = x , indep = y, dep_name = "expression", indep_name = "species" )

}
\keyword{ggplot}
\keyword{metabolomics}
