% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztransform.R
\name{ztransform}
\alias{ztransform}
\title{A copy of the ztransform function from the GenAbel package}
\usage{
ztransform(formula, data, family = gaussian)
}
\arguments{
\item{formula}{}

\item{data}{your vector of numerical values to rank normal transform}

\item{family}{Defaulted to gaussian. Do not suggest editing}
}
\description{
This function allows you to z  transform a vector of data.
}
\examples{
rntransform()
}
\keyword{nigtingale}
