% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_normalization.R
\name{batch_normalization}
\alias{batch_normalization}
\title{median batch normalization}
\usage{
batch_normalization(
  wdata,
  feature_data_sheet = NULL,
  sample_data_sheet = NULL,
  feature_runmode_col = NULL,
  batch_ids = NULL
)
}
\arguments{
\item{wdata}{the metabolite data frame samples in row, metabolites in columns}

\item{feature_data_sheet}{a data frame containing the feature annotation data}

\item{sample_data_sheet}{a data frame containing the sample annotation data}

\item{feature_runmode_col}{a string identifying the column name in the feature_data_sheet that identifies the run mode for each feature (metabolites of proteins).}

\item{batch_ids}{a string vector, with a length equal to the number of samples in the data set that identifies what batch each sample belongs to.}
}
\value{
returns the wdata object passed to the function median normalized given the batch information provided.
}
\description{
This function median normalizes multivariable data often processed in batches, such as metabolomic and proteomic data sets.
}
\examples{
####################################
## with a vector of batch variables
####################################
## define the data set
d1 = sapply(1:10, function(x){ rnorm(25, 40, 2) })
d2 = sapply(1:10, function(x){ rnorm(25, 35, 2) })
ex_data = rbind(d1,d2)
rownames(ex_data) = paste0("ind", 1:nrow(ex_data))
colnames(ex_data) = paste0("var", 1:ncol(ex_data))
## define the batch
batch = c( rep("A", 25), rep("B", 25)  )
## normalize by batch
norm_wdata = batch_normalization(wdata = ex_data, batch_ids = batch )

}
\keyword{batch}
\keyword{metabolomics}
\keyword{normalization}
\keyword{proteomics}
