% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.power.binary.R
\name{eval.power.binary}
\alias{eval.power.binary}
\title{Estimate power for a binary variable}
\usage{
eval.power.binary(N, effect, alpha)
}
\arguments{
\item{N}{a numeric vector of total study sample size, cases and controls will both be defined as N/2.}

\item{effect}{a numeric vector of effect size}

\item{alpha}{a numeric vector of significance thresholds}
}
\value{
a matrix of parameter inputs and an estimate(s) of power are returned as a matrix
}
\description{
This function allows you estimate power for a binary variable given the sample size, effect size, significance threshold.
}
\examples{
eval.power.binary(N = 1000, effect = seq(0.01, 0.3, by = 0.01), alpha = 0.05)

}
\keyword{binary}
\keyword{estimator}
\keyword{power}
