% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.power.cont.R
\name{eval.power.cont}
\alias{eval.power.cont}
\title{A Function to estimate power for a continuous variable}
\usage{
eval.power.cont(N, n_coeff, effect, alpha)
}
\arguments{
\item{N}{Sample size}

\item{n_coeff}{degrees of freedom for numerator}

\item{effect}{effect size}

\item{alpha}{significance level (Type 1 error)}
}
\description{
This function allows you estimate power for a continuous variable given the sample size, effect size, TypeI error, and the degrees of freedom. Useful for QC Rmd Report.
}
\examples{
eval.power.cont()
}
\keyword{metabolomics}
