% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.describe.R
\name{feature.describe}
\alias{feature.describe}
\title{A Function to describe each feature of a matrix of metabolite data.}
\usage{
feature.describe(wdata)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}
}
\description{
This function allows you to 'describe' metabolite features using the describe() function from the psych package, as well as estimate variance, a dispersion index, the coeficent of variation, and shapiro's W-statistic.
}
\examples{
feature.describe()
}
\keyword{metabolomics}
