% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.missingness.R
\name{feature.missingness}
\alias{feature.missingness}
\title{A Function to estiamte missingness on samples,}
\usage{
feature.missingness(wdata, samplemissingness)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{samplemissingness}{a vector of sample missingness for each sample}
}
\description{
This function allows you to estimate feature missingess, with a step to exclude poor samples identified as those with a sample missingness greater than 50%.
}
\examples{
feature.missingness()
}
\keyword{metabolomics}
