% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.missingness.R
\name{feature.missingness}
\alias{feature.missingness}
\title{estimate feature missingness}
\usage{
feature.missingness(wdata)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}
}
\value{
a data frame of percent missingness for each feature
}
\description{
This function estimates feature missingess, with a step to exclude poor samples identified as those with a sample missingness greater than 50%.
}
\examples{
ex_data = sapply(1:5, function(x){rnorm(10, 45, 2)})
ex_data[ sample(1:length(ex_data), 15) ] = NA
feature.missingness(wdata = ex_data )

}
\keyword{feature}
\keyword{missingness}
