% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.outliers.R
\name{feature.outliers}
\alias{feature.outliers}
\title{outlier sample count for a features}
\usage{
feature.outliers(wdata, nsd = 5)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{nsd}{the number of standard deviation from the mean outliers are identified at. Default value is 5.}
}
\value{
a data frame out sample outlier counts for each feature (column) in the matrix
}
\description{
This function takes a matrix of data (samples in rows, features in columns) and counts the number of outlying samples each feature has.
}
\examples{
ex_data = sapply(1:20, function(x){ rnorm(250, 40, 5) })
s = sample(1:length(ex_data), 200)
ex_data[s] = ex_data[s] + 40
## run the function
fout = feature.outliers(ex_data)

}
\keyword{feature}
\keyword{outliers}
