% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.sum.stats.R
\name{feature.sum.stats}
\alias{feature.sum.stats}
\title{A Function to estimate numerous feature summary statistics for your samples}
\usage{
feature.sum.stats(wdata, sammis = NA)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{sammis}{a vector of sample missingness estimates, that is ordered to match the samples in the rows of your data matrix.}
}
\description{
This function estiamtes feature statistics for samples in a matrix of metabolite features.
}
\examples{
feature.sum.stats()
}
\keyword{metabolomics}
