% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plots.R
\name{feature_plots}
\alias{feature_plots}
\title{A function to plot the distribution and a few summary statistics of each feature in a data frame to a pdf file}
\usage{
feature_plots(wdata, outdir = NULL, nsd = 5)
}
\arguments{
\item{wdata}{a data frame of feature (ex: metabolite or protein) abundance levels}

\item{outdir}{output directory path}

\item{nsd}{number of SD from the mean to plot an outlier line on the scatter plot and histogram}
}
\description{
This function to plots a scatter plot, a histogram, and a few summary statistics of each feature in a data frame to a pdf file
}
\examples{
feature_plots()
}
\keyword{metabolomics}
\keyword{pdf}
\keyword{summary}
