% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_summary.R
\name{feature_summary}
\alias{feature_summary}
\title{Feature Summary Statistics}
\usage{
feature_summary(
  metaboprep,
  source_layer = "input",
  outlier_udist = 5,
  tree_cut_height = 0.5,
  feature_ids = NULL,
  sample_ids = NULL,
  output = "data.frame"
)
}
\arguments{
\item{metaboprep}{an object of class Metabolites}

\item{source_layer}{character, the data layer to summarise}

\item{outlier_udist}{the unit distance in SD or IQR from the mean or median estimate, respectively outliers are identified at. Default value is 5.}

\item{tree_cut_height}{numeric, the threshold for feature independence in hierarchical clustering. Default is 0.5.}

\item{feature_ids}{character, vector of feature ids to work with}

\item{sample_ids}{character, vector of sample ids to work with}

\item{output}{character, type of output, either 'object' to return the updated metaboprep object, or 'data.frame' to return the data.}
}
\description{
This function estimates feature statistics for samples in a matrix of metabolite features.
}
