% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_summary.R
\name{feature_summary}
\alias{feature_summary}
\title{Feature Summary Statistics}
\usage{
feature_summary(
  metabolites,
  layer = "raw",
  feature_ids = NULL,
  sample_ids = NULL,
  output = "object"
)
}
\arguments{
\item{metabolites}{an object of class Metabolites}

\item{layer}{character, the data type/source}

\item{feature_ids}{character, vector of feature ids}

\item{sample_ids}{character, vector of sample ids}

\item{output}{character, the output format options: "object", "data.table", "matrix", "list"}
}
\description{
This function estimates feature statistics for samples in a matrix of metabolite features.
}
