% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.PA.effect.sizes.2.sim.R
\name{find.PA.effect.sizes.2.sim}
\alias{find.PA.effect.sizes.2.sim}
\title{A Function to estiamte an appropriate distribution of effect sizes to simulate in a power analysis.}
\usage{
find.PA.effect.sizes.2.sim(mydata)
}
\arguments{
\item{mydata}{Your metabolite data matrix, with samples in rows}
}
\description{
This function estiamtes an appropriate distribution of effect sizes to simulate in a power analysis.
}
\examples{
find.PA.effect.sizes.2.sim()
}
\keyword{metabolomics}
