% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{generate metaboprep summary html report}
\usage{
generate_report(
  full_path_2_Rdatafile = "ReportData.Rdata",
  dir_4_report = "./",
  path_2_Rmd_template = file.path(system.file("rmarkdown", package = "metaboprep"),
    "metaboprep_Report_v0.Rmd")
)
}
\arguments{
\item{full_path_2_Rdatafile}{full path to the Rdatafile}

\item{dir_4_report}{directory to place the report}

\item{path_2_Rmd_template}{full path to the html report template}
}
\value{
Writes a html report to file

an html file written to file
}
\description{
This function generates the html report.
}
\examples{


}
\keyword{html}
\keyword{knit}
\keyword{report}
