% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.pairwise.n.filter.R
\name{greedy.pairwise.n.filter}
\alias{greedy.pairwise.n.filter}
\title{A Function to identify features that have less than a minimum number of complete pairwise observations and remove one of them, in a greedy fashion.}
\usage{
greedy.pairwise.n.filter(wdata, minN = 50)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{minN}{the minimum sample size (n) for pairwise comparisons}
}
\description{
This function allows you to generate a dendrogram of feautres based on correlation coefficeint of your choice, and a clustering method of choice
}
\examples{
greedy.pairwise.n.filter()
}
\keyword{feature}
\keyword{selection}
