% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.outliers.R
\name{id.outliers}
\alias{id.outliers}
\title{A Function to identify SD outliers in a vector of data}
\usage{
id.outliers(x, nsd = 5)
}
\arguments{
\item{x}{a vector of numerical values}

\item{nsd}{the number of standard deviation from the mean outliers are identified at. Default value is 5.}
}
\description{
given a vector of data, identify those positions that are standard deviation outliers in that vector of data
}
\examples{
id.outliers()
}
\keyword{metabolomics}
