% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.cor.matrix.R
\name{make.cor.matrix}
\alias{make.cor.matrix}
\title{A fully bespoke function to generate a correaltion matrix of either correlation estimates or the pvalues}
\usage{
make.cor.matrix(wdata, cor_method = "kendall", minN = 50, var2return = "cor")
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{cor_method}{Defaulted to "kendall" this is the correlation method to use in the function cor.test()}

\item{minN}{Defaulted to 50, this is the minimum number of observations that must be available in pairs to perform analysis}

\item{var2return}{Defaulted to "cor", other option is "pvalue" is a the flag indicating which estimate to return from the function.}
}
\description{
This function allows estiamte sample missingness in a matrix of data and provides an option to exclude certain columns or metabolite features from the analysis, such as xenobiotics
}
\examples{
make.cor.matrix()
}
\keyword{nigtingale}
