% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.tree.R
\name{make.tree}
\alias{make.tree}
\title{A Function to generate a dendrogram of feautres based on correlation distances}
\usage{
make.tree(wdata, cor_method = "spearman", hclust_method = "complete")
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{cor_method}{the correlation method used in the function cor(). Default is "spearman".}

\item{hclust_method}{the dendrogram clustering method used in the construction of the tree. Default is "complete"}
}
\description{
This function allows you to generate a dendrogram of feautres based on correlation coefficeint of your choice, and a clustering method of choice
}
\examples{
make.tree()
}
\keyword{metabolomics}
