% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.matrix.R
\name{outlier.matrix}
\alias{outlier.matrix}
\title{A Function to identify outliers in each column of a matrix}
\usage{
outlier.matrix(wdata, nsd = 5)
}
\arguments{
\item{nsd}{the number of standard deviation from the mean outliers are identified at. Default value is 5.}

\item{x}{a vector of numerical values}
}
\description{
Given a matrix of data this function returns a matrix of 0|1, of the same structure with 1 values indicating outliers. It is an expansion of the function id.outliers(), applied to columns of a matrix.
}
\examples{
outlier.matrix()
}
\keyword{metabolomics}
