% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.summary.R
\name{outlier.summary}
\alias{outlier.summary}
\title{A Function to describe, plot the distribution of and identifiy outliers for every metabolite (column) in a data frame.}
\usage{
outlier.summary(dtst, pdf_filename = "./feature_distributions.pdf", nsd = 5)
}
\arguments{
\item{dtst}{metabolite data frame}

\item{pdf_filename}{name of the pdf out file}

\item{nsd}{number of SD to consider as outliers, 5 is default}
}
\description{
This function allows you to
}
\examples{
outlier.summary()
}
\keyword{metabolomics}
