% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{A Function to identify outliers from a vector of data at SD units from the mean.}
\usage{
outliers(x, nsd = 3)
}
\arguments{
\item{x}{a numerical vector of data}

\item{nsd}{the Number of SD units from the mean to be used as an outlier cutoff.}
}
\description{
This function to identify outliers from a vector of data at SD units from the mean.
}
\examples{
outliers()
}
\keyword{metabolomics}
