% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.and.outliers.R
\name{pc.and.outliers}
\alias{pc.and.outliers}
\title{A Function to estimate PC and then identify possible outliers}
\usage{
pc.and.outliers(metabolitedata, indfeature_names, outliers = TRUE)
}
\arguments{
\item{metabolitedata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{indfeature_names}{a vector of independent feature names | column names.}

\item{outliers}{Defaulted to TRUE, a TRUE|FALSE paramater flagging if you would like outliers identified.}
}
\description{
This function allows you to estiamte PCs and identify outliers in those PCs.
}
\examples{
pc.and.outliers()
}
\keyword{metabolomics}
