% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.factor.analysis.R
\name{pca.factor.analysis}
\alias{pca.factor.analysis}
\title{A Function to perform a PCA factor analysis.}
\usage{
pca.factor.analysis(metabolitedata, pcloadings, sigthreshold = 0.3,
  feature_anno = feature_data$SUPER_PATHWAY)
}
\arguments{
\item{metabolitedata}{data frame of metabolite data}

\item{pcloadings}{data frame of metabolite data}

\item{sigthreshold}{signficance threshold in spearman correlation units}

\item{feature_anno}{the feature annotation to include in the factor analysis}
}
\description{
This function allows you to perform a PCA factor analysis.
}
\examples{
pca.factor.analysis()
}
\keyword{metabolomics}
