% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcapairs_bymoose.R
\name{pcapairs_bymoose}
\alias{pcapairs_bymoose}
\title{A Function to plot PCs in a matrix}
\usage{
pcapairs_bymoose(myloadings, varexp)
}
\arguments{
\item{myloadings}{PC loadings}

\item{varexp}{the variance explained by each PC in a vector.}
}
\description{
This function to plot PCs in a matrix
}
\examples{
pcapairs_bymoose()
}
\keyword{metabolomics}
