% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{quality_control}
\alias{quality_control}
\title{Metabolite Quality Control}
\usage{
quality_control(metaboprep, source_layer = "input", dest_layer = "qc")
}
\arguments{
\item{metaboprep}{an object of class Metabolites}

\item{source_layer}{character, the data layer to summarise}
}
\description{
This function is a wrapper function that performs the key quality controls steps on a metabolomics data set.
Key principles:
 1. keep the source underlying data as it is
 2. copy the source data to a new data layer called qcing for processing
 3. build an exclusion list, accumulating codes for exclusion reasons
 4. make any adjustments needed in the destination copy of the data, flag these in the exclusion list
 5. copy the final result to a data layer called post_qc
 6. return the Metabolites object with the newly populated data layers
}
