% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.in.metabolon.R
\name{read.in.metabolon}
\alias{read.in.metabolon}
\title{read in Metabolon (v1) metabolomics data}
\usage{
read.in.metabolon(file2process, data_dir, projectname)
}
\arguments{
\item{file2process}{the name of the xls file to process}

\item{data_dir}{the full path to the directory holding your Metabolon excel file}

\item{projectname}{a name for your project}
}
\value{
a list object of (1) metabolite, (2) sample annotation, and (3) feature annotation data
}
\description{
This function reads in a Metabolon (v1 format) raw data excel file, writes the (1) metabolite, (2) sample annotation, and (3) feature annotation data to flat text files. It also returns a list object of the same data.
}
\examples{
read.in.metabolon(file2process = "Metabolon_data_release.xls", 
 data_dir = "/File/sits/here/", 
 projectname = "My Amazing Project")

}
\keyword{Metabolon}
\keyword{metabolomics}
