% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.in.nightingale.R
\name{read.in.nightingale}
\alias{read.in.nightingale}
\title{A Function to read in Nightingale excel sheets}
\usage{
read.in.nightingale(file2process, data_dir, projectname)
}
\arguments{
\item{file2process}{the name of the xls file to process}

\item{data_dir}{the full path to the directory holding your Nightingale excel file}

\item{projectname}{a name for your project}
}
\description{
This function allows you to read in Nightingale raw data and push out flat text files
}
\examples{
read.in.nightingale()
}
\keyword{nigtingale}
