% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metabolon.R
\name{read_metabolon}
\alias{read_metabolon}
\title{Read Metabolon Data}
\usage{
read_metabolon(filepath, sheet = NULL)
}
\arguments{
\item{filepath}{character, commercial Metabolon excel sheet with extension .xls or .xlsx}

\item{sheet}{character or integer, the excel sheet name (or index) from which to read.}
}
\value{
list,  list(data = matrix, samples = samples data.frame, features = features data.frame)
}
\description{
Read Metabolon Data
}
\examples{
# version 1 data format
filepath1 <- system.file("extdata", "metabolon_v1_example.xlsx", package = "metaboprep")
read_metabolon(filepath1, sheet = 2)

# version 2 data format
filepath2 <- system.file("extdata", "metabolon_v2_example.xlsx", package = "metaboprep")
read_metabolon(filepath2, sheet = 'OrigScale')

}
