% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nightingale.R
\name{read_nightingale}
\alias{read_nightingale}
\title{Read Nightingale Data (format 1)}
\usage{
read_nightingale(filepath)
}
\arguments{
\item{filepath}{character, commercial Nightingale excel sheet with extension .xls or .xlsx}
}
\value{
list,  list(data = 3D matrix, samples = samples data.table, features = features data.table)
}
\description{
Read Nightingale Data (format 1)
}
\examples{
# version 1 data format
filepath1 <- system.file("extdata", "nightingale_v1_example.xlsx", package = "metaboprep")
read_nightingale(filepath1)

# version 2 data format
filepath2 <- system.file("extdata", "nightingale_v2_example.xlsx", package = "metaboprep")
read_nightingale(filepath2)

}
