% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nightingale.R
\name{read_nightingale}
\alias{read_nightingale}
\title{Read Nightingale Data (format 1)}
\usage{
read_nightingale(filepath, return_Metaboprep = TRUE)
}
\arguments{
\item{filepath}{character, commercial Nightingale excel sheet with extension .xls or .xlsx}

\item{return_Metaboprep}{logical, if TRUE (default) return a Metaboprep object, if FALSE return a list.}
}
\value{
list or Metaboprep object, list(data = matrix, samples = samples data.frame, features = features data.frame)
}
\description{
Read Nightingale Data (format 1)
}
\examples{
# version 1 data format
filepath1 <- system.file("extdata", "nightingale_v1_example.xlsx", package = "metaboprep")
m <- read_nightingale(filepath1)

# version 2 data format
filepath2 <- system.file("extdata", "nightingale_v2_example.xlsx", package = "metaboprep")
m <- read_nightingale(filepath2)

}
