% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_somalogic.R
\name{read_somalogic}
\alias{read_somalogic}
\title{Read Somalogic Data}
\usage{
read_somalogic(filepath)
}
\arguments{
\item{filepath}{character, commercial Somalogicfile}
}
\value{
list,  list(data = 3D matrix, samples = samples data.table, features = features data.table)
}
\description{
Read Somalogic Data
}
