% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.outliers.R
\name{sample.outliers}
\alias{sample.outliers}
\title{outlier features count for samples}
\usage{
sample.outliers(wdata, nsd = 5)
}
\arguments{
\item{wdata}{a metabolite data matrix with samples in row, metabolites in columns}

\item{nsd}{the number of standard deviation from the mean outliers are identified at. The default value is 5.}
}
\value{
a data frame of outiler counts for each sample
}
\description{
This function takes a matrix of numeric data and counts the number of outlying features each sample has.
}
\examples{
d = sapply(1:5, function(x){ rnorm(50, 50, 15) })
sample.outliers(d, nsd = 2)

}
\keyword{metabolomics}
\keyword{outliers}
