% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.sum.stats.R
\name{sample.sum.stats}
\alias{sample.sum.stats}
\title{summary statistics for samples}
\usage{
sample.sum.stats(wdata, feature_names_2_exclude = NA, outlier_udist = 5)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}

\item{feature_names_2_exclude}{a vector of feature|column names to exclude from missingness estimates}

\item{outlier_udist}{the interquartile range unit distance from the median to call a sample an outlier at a feature.}
}
\value{
a data frame of summary statistics
}
\description{
This function estimates summary statistics for samples in a matrix of numeric features. This includes missingness, total peak area, and a count of the number of outlying features for a sample.
}
\examples{
## simulate some data
set.seed(1110)
ex_data = sapply(1:5, function(x){ rnorm(10, 40, 5) })
rownames(ex_data) = paste0("ind", 1:nrow(ex_data))
colnames(ex_data) = paste0("var", 1:ncol(ex_data))
## add some missingness to the data
ex_data[ sample(1:50, 10) ] = NA
## run estimate sample summary statistics
sample.sum.stats(ex_data)

}
\keyword{sample}
\keyword{statistics}
\keyword{summary}
