% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.peak.area.R
\name{total.peak.area}
\alias{total.peak.area}
\title{A Function estiamtes total peak area for each sample of a metabolite matrix}
\usage{
total.peak.area(wdata)
}
\arguments{
\item{wdata}{the metabolite data matrix. samples in row, metabolites in columns}
}
\description{
This function estimates total peak area for all features and all features with complete data.
}
\examples{
total.peak.area()
}
\keyword{metabolomics}
